ADD: doxygen documentation
